+++ src/main/java/kitchenpos/product/application/port/in/ProductFindUseCase.java
@@ -1,0 +1,9 @@
+package kitchenpos.product.application.port.in;
+
+import java.util.List;
+import kitchenpos.product.application.ProductDTO;
+
+public interface ProductFindUseCase {
+
+    List<ProductDTO> findAll();
+}
